+++ app/api/questions.py
@@ -67,10 +67,11 @@ async def create_question(
 @router.put("/{question_id}", status_code=status.HTTP_200_OK, dependencies=[Depends(auth_user_jwt)])
 async def update_question(
     question_id: str,
-    data: dict,
+    request: Request,
     question_service: Annotated[QuestionManager, Depends(get_question_service)]
 ) -> Question:
     try:
+        data = await request.json()
         return await question_service.update_question(data, question_id)
     except ValidationError as e:
         raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail=str(e))