+++ js/components/auth/main.js
@@ -6,5 +6,5 @@ import Page from "../common/page.jsx";
 document.addEventListener("DOMContentLoaded", () => {
   const container = document.getElementById("authDiv");
   const root = createRoot(container);
-  root.render(React.createElement(Page, { pageTitle: "Authenticate", content: Authenticate, openSidebar: false }));
+  root.render(React.createElement(Page, { pageTitle: "Admin Login", content: Authenticate, openSidebar: false }));
 });